+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+1.12.0 / 2018-03-25
+===================
+
+* Browser.assert.tag and browser.assert.not.tag
+* Browser.tag
+* Browser.findByAttribute
+* Added proxyServer option in create browser
+* Minor fixes in some error messages
+
 1.11.0 / 2018-03-24
 ===================